+++ src/command_modules/azure-cli-vm/azure/cli/command_modules/vm/_template_builder.py
@@ -250,8 +250,9 @@ def build_vm_resource(  # pylint: disable=too-many-locals
         disk_info=None, boot_diagnostics_storage_uri=None, ultra_ssd_enabled=None):
 
     os_caching = disk_info['os'].get('caching')
-    # TODO: split out the storage_sku for os disk(can't be ultra-ssd) and individual data disks
-    os_storage_sku = None if storage_sku == 'UltraSSD_LRS' else storage_sku
+    os_storage_sku = disk_info['os'].get('storageAccountType')
+    if os_storage_sku and os_storage_sku.lower() == 'UltraSSD_LRS'.lower():
+        os_storage_sku = None
 
     def _build_os_profile():
 
@@ -338,7 +339,7 @@ def build_vm_resource(  # pylint: disable=too-many-locals
                     'createOption': 'fromImage',
                     'name': os_disk_name,
                     'caching': os_caching,
-                    'managedDisk': {'storageAccountType': storage_sku}
+                    'managedDisk': {'storageAccountType': os_storage_sku}
                 },
                 "imageReference": {
                     'id': image_reference
@@ -614,8 +615,8 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
                         vm_sku, instance_count, ip_config_name, nic_name, subnet_id,
                         public_ip_per_vm, vm_domain_name, dns_servers, nsg, accelerated_networking,
                         admin_username, authentication_type, storage_profile, os_disk_name, disk_info,
-                        storage_sku, os_type, image=None, admin_password=None, ssh_key_value=None, ssh_key_path=None,
-                        os_publisher=None, os_offer=None, os_sku=None, os_version=None,
+                        os_type, storage_sku=None, image=None, admin_password=None, ssh_key_value=None,
+                        ssh_key_path=None, os_publisher=None, os_offer=None, os_sku=None, os_version=None,
                         backend_address_pool_id=None, inbound_nat_pool_id=None, health_probe=None,
                         single_placement_group=None, platform_fault_domain_count=None, custom_data=None,
                         secrets=None, license_type=None, zones=None, priority=None, eviction_policy=None,
@@ -660,8 +661,10 @@ def build_vmss_resource(cmd, name, naming_prefix, location, tags, overprovision,
     # Build storage profile
     storage_properties = {}
     os_caching = disk_info['os'].get('caching')
-    # TODO: split out the storage_sku for os disk(can't be ultra-ssd) and individual data disks
-    os_storage_sku = None if storage_sku == 'UltraSSD_LRS' else storage_sku
+
+    os_storage_sku = disk_info['os'].get('storageAccountType')
+    if os_storage_sku and os_storage_sku.lower() == 'UltraSSD_LRS'.lower():
+        os_storage_sku = None
 
     if storage_profile in [StorageProfile.SACustomImage, StorageProfile.SAPirImage]:
         storage_properties['osDisk'] = {