+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -10,6 +10,7 @@ Release History
 * `vm encryption show`: table output format shows status
 * `vm secret format`: requires json/jsonc output. Warns user and defaults to json output if an undesired output format is selected.
 * `vm create --image`: improved validation of image argument
+* `vm create --storage-sku`: can now specify the storage account sku for managed os and data disks separately.
 
 2.2.6
 ++++++